requirements.txt
absl-py==2.1.0
altgraph==0.17.4
arrow==1.3.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
fastkml==1.1.0
filelock==3.16.1
fiona==1.10.1
fsspec==2025.3.0
geopandas==0.13.2
google-auth==2.40.1
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.70.0
huggingface-hub==0.31.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.6
joblib==1.4.2
kismet-analyzer==0.4.1
lxml==5.4.0
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==25.0
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pygeoif==1.5.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyproj==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
shapely==2.0.7
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.6
zipp==3.20.2
